requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
openai>=1.0.0
httpx>=0.24.0
python-dotenv>=0.19.0
tenacity>=8.0.0
requests>=2.28.0
aiohttp>=3.8.0
sentence_transformers
faiss-cpu
PyPDF2
python-docx
python-multipart>=0.0.5
loguru==0.7.2
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.1.1
cryptography  
python-jose[cryptography]>=3.3.0
passlib==1.7.4
bcrypt==4.0.1
email-validator>=1.1.3
psutil>=7.0.0
GPUtil>=1.4.0
redis>=5.0.0
langgraph==0.3.25
langchain-deepseek==0.1.3
langchain_neo4j==0.4.0
matplotlib==3.10.1
seaborn==0.13.2
langchain_community==0.3.21
# graphrag==2.1.0
langgraph-checkpoint-sqlite==2.0.6
ipython==9.1.0

# Wechat MP Server Dependencies
Flask>=3.0.0
wx-mp-svr
wx-crypt

# GraphRAG Dependencies
environs>=11.0.0
azure-search-documents>=11.5.2
lancedb>=0.17.0
aiofiles>=24.1.0
fnllm>=0.2.3
json-repair>=0.30.3
nltk>=3.9.1
tiktoken>=0.8.0
numpy>=1.25.2,<2.0.0

networkx>=3.4.2
pandas>=2.2.3
pyarrow>=15.0.0
umap-learn>=0.5.6
pyyaml>=6.0.2
rich>=13.9.4
devtools>=0.12.2
azure-cosmos>=4.9.0
azure-identity>=1.19.0
azure-storage-blob>=12.24.0
future>=1.0.0
typer>=0.15.1
tqdm>=4.67.1
textblob>=0.18.0.post0
spacy>=3.8.4